requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
joblib==1.3.2 